requirements.txt
python-telegram-bot[job-queue]==21.6
reportlab==4.2.5
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
